+++ smart-contracts/contracts/mixins/MixinLockCore.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-// import '@openzeppelin/contracts-upgradeable/token/ERC721/extensions/IERC721EnumerableUpgradeable.sol';
 import '@openzeppelin/contracts-upgradeable/utils/AddressUpgradeable.sol';
 import './MixinDisable.sol';
 import './MixinRoles.sol';
@@ -13,10 +12,10 @@ import '../interfaces/hooks/ILockKeyPurchaseHook.sol';
 import '../interfaces/hooks/ILockValidKeyHook.sol';
 import '../interfaces/hooks/ILockTokenURIHook.sol';
 import '../interfaces/hooks/ILockKeyTransferHook.sol';
+import '../interfaces/hooks/ILockKeyExtendHook.sol';
 
 /**
  * @title Mixin for core lock data and functions.
- * @author HardlyDifficult
  * @dev `Mixins` are a design pattern seen in the 0x contracts.  It simply
  * separates logically groupings of code to ease readability.
  */
@@ -30,7 +29,7 @@ contract MixinLockCore is
   event Withdrawal(
     address indexed sender,
     address indexed tokenAddress,
-    address indexed beneficiary,
+    address indexed recipient,
     uint amount
   );
 
@@ -57,7 +56,6 @@ contract MixinLockCore is
   event ApprovalForAll(address indexed owner, address indexed operator, bool approved);
 
   // Unlock Protocol address
-  // TODO: should we make that private/internal?
   IUnlock public unlockProtocol;
 
   // Duration in seconds for which the keys are valid, after creation
@@ -65,7 +63,6 @@ contract MixinLockCore is
   uint public expirationDuration;
 
   // price in wei of the next key
-  // TODO: allow support for a keyPriceCalculator which could set prices dynamically
   uint public keyPrice;
 
   // Max number of keys sold if the keyReleaseMechanism is public
@@ -74,8 +71,8 @@ contract MixinLockCore is
   // A count of how many new key purchases there have been
   uint internal _totalSupply;
 
-  // The account which will receive funds on withdrawal
-  address payable public beneficiary;
+  // DEPREC: this is not used anymore (kept as private var for storage layout compat)
+  address private beneficiary;
 
   // The denominator component for values specified in basis points.
   uint internal constant BASIS_POINTS_DEN = 10000;
@@ -95,32 +92,24 @@ contract MixinLockCore is
   // one more hook (added to v11)
   ILockKeyTransferHook public onKeyTransferHook;
 
+  // one more hook (added to v12)
+  ILockKeyExtendHook public onKeyExtendHook;
+
   // modifier to check if data has been upgraded
   function _lockIsUpToDate() internal view {
     if(schemaVersion != publicLockVersion()) {
       revert MIGRATION_REQUIRED();
     }
   }
-
-  // modifier
-  function _onlyLockManagerOrBeneficiary() 
-  internal 
-  view
-  {
-    if(!isLockManager(msg.sender) && msg.sender != beneficiary) {
-      revert ONLY_LOCK_MANAGER_OR_BENEFICIARY();
-    }
-  }
   
   function _initializeMixinLockCore(
-    address payable _beneficiary,
+    address payable _lockOwner,
     uint _expirationDuration,
     uint _keyPrice,
     uint _maxNumberOfKeys
   ) internal
   {
     unlockProtocol = IUnlock(msg.sender); // Make sure we link back to Unlock's smart contract.
-    beneficiary = _beneficiary;
     expirationDuration = _expirationDuration;
     keyPrice = _keyPrice;
     maxNumberOfKeys = _maxNumberOfKeys;
@@ -137,22 +126,22 @@ contract MixinLockCore is
   ) public pure
     returns (uint16)
   {
-    return 11;
+    return 12;
   }
 
   /**
-   * @dev Called by owner to withdraw all funds from the lock and send them to the `beneficiary`.
-   * @param _tokenAddress specifies the token address to withdraw or 0 for ETH. This is usually
-   * the same as `tokenAddress` in MixinFunds.
+   * @dev Called by owner to withdraw all ETH funds from the lock
+   * @param _recipient specifies the address to send ETH to.
    * @param _amount specifies the max amount to withdraw, which may be reduced when
    * considering the available balance. Set to 0 or MAX_UINT to withdraw everything.
    */
   function withdraw(
     address _tokenAddress,
+    address payable _recipient,
     uint _amount
   ) external
   {
-    _onlyLockManagerOrBeneficiary();
+    _onlyLockManager();
 
     // get balance
     uint balance;
@@ -175,9 +164,9 @@ contract MixinLockCore is
       amount = _amount;
     }
 
-    emit Withdrawal(msg.sender, _tokenAddress, beneficiary, amount);
+    emit Withdrawal(msg.sender, _tokenAddress, _recipient, amount);
     // Security: re-entrancy not a risk as this is the last line of an external function
-    _transfer(_tokenAddress, beneficiary, amount);
+    _transfer(_tokenAddress, _recipient, amount);
   }
 
   /**
@@ -201,20 +190,6 @@ contract MixinLockCore is
   }
 
   /**
-   * A function which lets the owner of the lock update the beneficiary account,
-   * which receives funds on withdrawal.
-   */
-  function updateBeneficiary(
-    address payable _beneficiary
-  ) external {
-    _onlyLockManagerOrBeneficiary();
-    if(_beneficiary == address(0)) {
-      revert INVALID_ADDRESS();
-    }
-    beneficiary = _beneficiary;
-  }
-
-  /**
    * @notice Allows a lock manager to add or remove an event hook
    */
   function setEventHooks(
@@ -222,7 +197,8 @@ contract MixinLockCore is
     address _onKeyCancelHook,
     address _onValidKeyHook,
     address _onTokenURIHook,
-    address _onKeyTransferHook
+    address _onKeyTransferHook,
+    address _onKeyExtendHook
   ) external
   {
     _onlyLockManager();
@@ -232,12 +208,14 @@ contract MixinLockCore is
     if(_onValidKeyHook != address(0) && !_onValidKeyHook.isContract()) { revert INVALID_HOOK(2); }
     if(_onTokenURIHook != address(0) && !_onTokenURIHook.isContract()) { revert INVALID_HOOK(3); }
     if(_onKeyTransferHook != address(0) && !_onKeyTransferHook.isContract()) { revert INVALID_HOOK(4); }
+    if(_onKeyExtendHook != address(0) && !_onKeyExtendHook.isContract()) { revert INVALID_HOOK(5); }
     
     onKeyPurchaseHook = ILockKeyPurchaseHook(_onKeyPurchaseHook);
     onKeyCancelHook = ILockKeyCancelHook(_onKeyCancelHook);
     onTokenURIHook = ILockTokenURIHook(_onTokenURIHook);
     onValidKeyHook = ILockValidKeyHook(_onValidKeyHook);
     onKeyTransferHook = ILockKeyTransferHook(_onKeyTransferHook);
+    onKeyExtendHook = ILockKeyExtendHook(_onKeyExtendHook);
   }
 
   function totalSupply()
@@ -247,23 +225,8 @@ contract MixinLockCore is
     return _totalSupply;
   }
 
-  /**
-   * @notice An ERC-20 style approval, allowing the spender to transfer funds directly from this lock.
-   * @param _spender address that can spend tokens belonging to the lock
-   * @param _amount amount of tokens that can be spent by the spender
-   */
-  function approveBeneficiary(
-    address _spender,
-    uint _amount
-  ) public
-    returns (bool)
-  {
-    _onlyLockManagerOrBeneficiary();
-    return IERC20Upgradeable(tokenAddress).approve(_spender, _amount);
-  }
-
-
   // decreased from 1000 to 998 when adding `schemaVersion` and `maxKeysPerAddress` in v10 
   // decreased from 998 to 997 when adding `onKeyTransferHook` in v11
-  uint256[997] private __safe_upgrade_gap;
+  // decreased from 997 to 996 when adding `onKeyExtendHook` in v11
+  uint256[996] private __safe_upgrade_gap;
 }