+++ unlock-app/src/components/interface/locks/Create/modals/SelectCurrencyModal.tsx
@@ -2,12 +2,12 @@ import { Dialog, Transition } from '@headlessui/react'
 import { Token } from '@unlock-protocol/types'
 import { Button, Input } from '@unlock-protocol/ui'
 import { Fragment, useEffect, useState } from 'react'
-import useDebounce from '~/hooks/useDebouce'
-import { utils } from 'ethers'
+import { useDebounce } from 'react-use'
+import { ethers, utils } from 'ethers'
 import { useConfig } from '~/utils/withConfig'
 import { CryptoIcon } from '../../elements/KeyPrice'
 import { addressMinify } from '~/utils/strings'
-import { useQuery } from 'react-query'
+import { useQuery } from '@tanstack/react-query'
 import { useWeb3Service } from '~/utils/withWeb3Service'
 import { FaSpinner as Spinner } from 'react-icons/fa'
 import { useForm } from 'react-hook-form'
@@ -19,6 +19,8 @@ interface SelectCurrencyModalProps {
   defaultCurrency: string
 }
 
+export const ZERO = ethers.constants.AddressZero
+
 export const SelectCurrencyModal = ({
   isOpen,
   setIsOpen,
@@ -29,8 +31,21 @@ export const SelectCurrencyModal = ({
   const { networks } = useConfig()
   const web3Service = useWeb3Service()
   const [contractAddress, setContractAddress] = useState<string>('')
-  const [query, setQuery] = useState<string>('')
-  const queryValue = useDebounce<string>(query)
+  const [query, setQuery] = useState('')
+  const [_isReady] = useDebounce(
+    () => {
+      setQuery(query)
+      try {
+        const address = utils.getAddress(query)
+        setContractAddress(address)
+      } catch (err: any) {
+        setContractAddress('')
+        console.error('Error: ', err)
+      }
+    },
+    500,
+    [query]
+  )
   const { tokens: tokenItems = [] } = networks[network!] || {}
   const [tokens, setTokens] = useState<Token[]>([])
 
@@ -43,7 +58,7 @@ export const SelectCurrencyModal = ({
 
   useEffect(() => {
     setTokens([
-      { name: defaultCurrency, symbol: defaultCurrency },
+      { name: defaultCurrency, symbol: defaultCurrency, address: ZERO },
       ...tokenItems,
     ])
   }, [network])
@@ -55,26 +70,10 @@ export const SelectCurrencyModal = ({
     }
   }
 
-  const onSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e?.target?.value || ''
-    setQuery(value)
-  }
-
-  useEffect(() => {
-    if (!queryValue?.length) return
-    try {
-      const address = utils.getAddress(queryValue)
-      setContractAddress(address)
-    } catch (err: any) {
-      setContractAddress('')
-      console.error('Error: ', err)
-    }
-  }, [queryValue])
-
   const tokensFiltered = tokens?.filter(
     (token: Token) =>
-      token?.name?.toLowerCase().includes(queryValue?.toLowerCase()) ||
-      token?.symbol?.toLowerCase().includes(queryValue?.toLowerCase())
+      token?.name?.toLowerCase().includes(query?.toLowerCase()) ||
+      token?.symbol?.toLowerCase().includes(query?.toLowerCase())
   )
 
   const getContractTokenSymbol = async () => {
@@ -82,15 +81,14 @@ export const SelectCurrencyModal = ({
   }
 
   const { isLoading: isLoadingContractToken, data: contractTokenSymbol } =
-    useQuery(
-      ['getContractTokenSymbol', contractAddress, queryValue],
-      async () => getContractTokenSymbol()
+    useQuery(['getContractTokenSymbol', contractAddress, query], async () =>
+      getContractTokenSymbol()
     )
 
   const addToken = ({
     name,
     symbol,
-    address = undefined,
+    address,
     decimals = 18,
   }: Partial<Token>) => {
     const currentList = tokens || []
@@ -159,7 +157,7 @@ export const SelectCurrencyModal = ({
                     className="bg-transparent"
                     autoComplete="off"
                     {...register('query', {
-                      onChange: onSearch,
+                      onChange: (e) => setQuery(e.target.value),
                     })}
                   />