+++ unlock-app/src/components/interface/locks/List/elements/LockCard.tsx
@@ -2,7 +2,7 @@ import { addressMinify } from '~/utils/strings'
 import { useConfig } from '~/utils/withConfig'
 import useClipboard from 'react-use-clipboard'
 import { ToastHelper } from '~/components/helpers/toast.helper'
-import { Button, Icon } from '@unlock-protocol/ui'
+import { Button, Icon, Tooltip } from '@unlock-protocol/ui'
 import React, { useEffect, useState } from 'react'
 import { BiCopy as CopyIcon } from 'react-icons/bi'
 import { HiOutlineExternalLink as ExternalLinkIcon } from 'react-icons/hi'
@@ -44,8 +44,10 @@ const Detail = ({ label, prepend, icon, value = '-' }: DetailProps) => {
         <span className="text-xs">{label}</span>
       </div>
       <div className="flex items-center gap-2">
-        {prepend && <>{prepend}</>}
-        <span className="text-lg font-bold">{value}</span>
+        {prepend && <div>{prepend}</div>}
+        <Tooltip tip={value} label={label} side="bottom">
+          <span className="text-lg font-bold truncate">{value}</span>
+        </Tooltip>
       </div>
     </div>
   )
@@ -195,25 +197,25 @@ export const LockCard = ({ lock, network, isLoading }: LockCardProps) => {
               </div>
             </div>
           </div>
-          <div className="grid items-center grid-cols-3 gap-3 md:grid-cols-4 md:col-span-3 md:gap-14">
-            <div className="col-span-1">
-              <Detail
-                label="Price"
-                value={lock.keyPrice}
-                icon={TagIcon}
-                prepend={<CryptoIcon symbol={symbol} size={25} />}
-              />
-            </div>
-            <div className="col-span-1 md:col-span-2">
-              <Detail label="Key Duration" value={duration} icon={TimeIcon} />
-            </div>
-            <div className="col-span-1">
-              <Detail
-                label="Key Sold"
-                value={lock.outstandingKeys?.toString()}
-                icon={KeyIcon}
-              />
-            </div>
+          <div className="grid items-center grid-cols-4 gap-3 md:col-span-3 md:gap-14">
+            <Detail
+              label="Price"
+              value={lock.keyPrice}
+              icon={TagIcon}
+              prepend={<CryptoIcon symbol={symbol} size={25} />}
+            />
+            <Detail
+              label="Balance"
+              value={lock.balance}
+              icon={TagIcon}
+              prepend={<CryptoIcon symbol={symbol} size={25} />}
+            />
+            <Detail label="Key Duration" value={duration} icon={TimeIcon} />
+            <Detail
+              label="Key Sold"
+              value={lock.outstandingKeys?.toString()}
+              icon={KeyIcon}
+            />
           </div>
           <div className="md:ml-auto md:col-span-1">
             <Link href={lockUrl} aria-label="arrow right">